+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Link to `pomoxis` in `README.md`.
+### Changed
+- The program will now exit with a zero exit code when errors are encountered in parsing input files or records.
+  The previous behaviour was to exit with a non-zero exit code, and can be restored with the `--force_error` option.
+- The per-file output has been amended to include a set of columns detailing failures during file parsing, and the results of read length, quality, and dust filtering.
+
 
 ## [v0.22.1]
 ### Added